+++ src/argilla/client/feedback/dataset/remote/dataset.py
@@ -105,6 +105,14 @@ class RemoteFeedbackRecords(ArgillaRecordsMixin):
         """Returns the `httpx.Client` instance that will be used to send requests to Argilla."""
         return self.dataset._client
 
+    @property
+    def _question_id_to_name(self) -> Dict["UUID", str]:
+        return self.dataset._question_id_to_name
+
+    @property
+    def _question_name_to_id(self) -> Dict[str, "UUID"]:
+        return self.dataset._question_name_to_id
+
     @allowed_for_roles(roles=[UserRole.owner, UserRole.admin])
     def __len__(self) -> int:
         """Returns the number of records in the current `FeedbackDataset` in Argilla."""
@@ -240,7 +248,7 @@ class RemoteFeedbackRecords(ArgillaRecordsMixin):
         )
 
 
-class RemoteFeedbackDataset(FeedbackDatasetBase):
+class RemoteFeedbackDataset(FeedbackDatasetBase[RemoteFeedbackRecord]):
     # TODO: Call super method once the base init contains only commons init attributes
     def __init__(
         self,
@@ -311,6 +319,18 @@ class RemoteFeedbackDataset(FeedbackDatasetBase):
         """
         return self._records
 
+    def update_records(self, records: Union[RemoteFeedbackRecord, List[RemoteFeedbackRecord]]) -> None:
+        if not isinstance(records, list):
+            records = [records]
+
+        metadata_schema = self._build_metadata_schema()
+        for record in records:
+            self._validate_record_metadata(record=record, metadata_schema=metadata_schema)
+
+        # TODO: Use the batch version of endpoint once is implemented
+        for record in records:
+            record.update()
+
     @property
     def id(self) -> "UUID":
         """Returns the ID of the dataset in Argilla."""
@@ -340,6 +360,14 @@ class RemoteFeedbackDataset(FeedbackDatasetBase):
     def updated_at(self) -> datetime:
         """Returns the datetime when the dataset was last updated in Argilla."""
         return self._updated_at
+
+    @property
+    def _question_id_to_name(self) -> Dict["UUID", str]:
+        return {question.id: question.name for question in self._questions}
+
+    @property
+    def _question_name_to_id(self) -> Dict[str, "UUID"]:
+        return {question.name: question.id for question in self._questions}
 
     def __repr__(self) -> str:
         """Returns a string representation of the dataset."""