+++ src/argilla/server/schemas/v1/datasets.py
@@ -24,6 +24,7 @@ from pydantic.utils import GetterDict
 
 from argilla.server.enums import UserRole
 from argilla.server.schemas.base import UpdateSchema
+from argilla.server.schemas.v1.records import RecordUpdate
 from argilla.server.schemas.v1.suggestions import Suggestion, SuggestionCreate
 from argilla.server.search_engine import StringQuery
 
@@ -57,9 +58,9 @@ QUESTION_CREATE_DESCRIPTION_MAX_LENGTH = 1000
 
 METADATA_PROPERTY_CREATE_NAME_REGEX = r"^(?=.*[a-z0-9])[a-z0-9_-]+$"
 METADATA_PROPERTY_CREATE_NAME_MIN_LENGTH = 1
-METADATA_PROPERTY_CREATE_NAME_MAX_LENGTH = 50
-METADATA_PROPERTY_CREATE_DESCRIPTION_MIN_LENGTH = 1
-METADATA_PROPERTY_CREATE_DESCRIPTION_MAX_LENGTH = 1000
+METADATA_PROPERTY_CREATE_NAME_MAX_LENGTH = 200
+METADATA_PROPERTY_CREATE_TITLE_MIN_LENGTH = 1
+METADATA_PROPERTY_CREATE_TITLE_MAX_LENGTH = 500
 
 RATING_OPTIONS_MIN_ITEMS = 2
 RATING_OPTIONS_MAX_ITEMS = 10
@@ -84,6 +85,9 @@ TERMS_METADATA_PROPERTY_MIN_VALUES = 1
 
 RECORDS_CREATE_MIN_ITEMS = 1
 RECORDS_CREATE_MAX_ITEMS = 1000
+
+RECORDS_UPDATE_MIN_ITEMS = 1
+RECORDS_UPDATE_MAX_ITEMS = 1000
 
 
 class Dataset(BaseModel):
@@ -357,9 +361,9 @@ class RecordGetterDict(GetterDict):
     def get(self, key: str, default: Any) -> Any:
         if key == "metadata":
             return getattr(self._obj, "metadata_", None)
-        if key == "responses" and "responses" not in self._obj.__dict__:
+        if key == "responses" and not self._obj.is_relationship_loaded("responses"):
             return default
-        if key == "suggestions" and "suggestions" not in self._obj.__dict__:
+        if key == "suggestions" and not self._obj.is_relationship_loaded("suggestions"):
             return default
         return super().get(key, default)
 
@@ -432,6 +436,16 @@ class RecordsCreate(BaseModel):
     )
 
 
+class RecordUpdateWithId(RecordUpdate):
+    id: UUID
+
+
+class RecordsUpdate(BaseModel):
+    items: List[RecordUpdateWithId] = PydanticField(
+        ..., min_items=RECORDS_UPDATE_MIN_ITEMS, max_items=RECORDS_UPDATE_MAX_ITEMS
+    )
+
+
 NT = TypeVar("NT", int, float)
 
 
@@ -463,6 +477,11 @@ class FloatMetadataPropertyCreate(NumericMetadataProperty[float]):
     type: Literal[MetadataPropertyType.float]
 
 
+MetadataPropertyTitleCreate = Annotated[
+    constr(min_length=METADATA_PROPERTY_CREATE_TITLE_MIN_LENGTH, max_length=METADATA_PROPERTY_CREATE_TITLE_MAX_LENGTH),
+    PydanticField(..., description="The title of the metadata property"),
+]
+
 MetadataPropertySettingsCreate = Annotated[
     Union[TermsMetadataPropertyCreate, IntegerMetadataPropertyCreate, FloatMetadataPropertyCreate],
     PydanticField(..., discriminator="type"),
@@ -476,11 +495,7 @@ class MetadataPropertyCreate(BaseModel):
         min_length=METADATA_PROPERTY_CREATE_NAME_MIN_LENGTH,
         max_length=METADATA_PROPERTY_CREATE_NAME_MAX_LENGTH,
     )
-    description: Optional[str] = PydanticField(
-        None,
-        min_length=METADATA_PROPERTY_CREATE_DESCRIPTION_MIN_LENGTH,
-        max_length=METADATA_PROPERTY_CREATE_DESCRIPTION_MAX_LENGTH,
-    )
+    title: MetadataPropertyTitleCreate
     settings: MetadataPropertySettingsCreate
     visible_for_annotators: bool = True
 
@@ -511,7 +526,7 @@ MetadataPropertySettings = Annotated[
 class MetadataProperty(BaseModel):
     id: UUID
     name: str
-    description: Optional[str] = None
+    title: str
     settings: MetadataPropertySettings
     visible_for_annotators: bool
     inserted_at: datetime